firestore adding and fetching plus filtering
